pyproject.toml
[tool.poetry]
name = "x post bot"
version = "0.1.0"
description = ""
authors = ["clement.machado <cm@helloai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.14.0"
python-dotenv = "^1.0.1"
openai = "^1.51.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
